requirements.txt
aiohttp==3.8.1
aiosignal==1.2.0
async-timeout==4.0.2
attrs==22.1.0
boto3==1.24.71
botocore==1.27.71
certifi==2022.6.15.1
charset-normalizer==2.1.1
click==8.1.3
datasets==2.4.0
diffusers==0.2.4
dill==0.3.5.1
fairscale==0.4.4
filelock==3.8.0
frozenlist==1.3.1
fsspec==2022.8.2
ftfy==6.1.1
huggingface-hub==0.9.1
idna==3.3
importlib-metadata==4.12.0
jmespath==1.0.1
joblib==1.1.0
multidict==6.0.2
multiprocess==0.70.13
numpy==1.21.6
packaging==21.3
pandas==1.3.5
Pillow==9.2.0
pyarrow==9.0.0
pyparsing==3.0.9
python-dateutil==2.8.2
pytz==2022.2.1
PyYAML==6.0
regex==2022.9.13
requests==2.28.1
responses==0.18.0
s3transfer==0.6.0
sacremoses==0.0.53
scipy==1.7.3
sentencepiece==0.1.97
six==1.16.0
timm==0.4.12
tokenizers==0.12.1
torch==1.12.1
tqdm==4.64.1
transformers==4.21.3
typing_extensions==4.3.0
urllib3==1.26.12
wandb==0.13.3
wcwidth==0.2.5
xxhash==3.0.0
yapf==0.32.0
yarl==1.8.1
zipp==3.8.1